# Add basic performance benchmarks for binary parsing