This adds a simple suite of performance benchmarks for parsing from the binary format.

`make bench` outputs:

```
### Parsing binary
large google.protobuf.FileDescriptorSet (1027010 bytes) x 29.20 ops/sec ±6.57% (51 runs sampled)
tiny docs.User (4 bytes) x 2,482,014 ops/sec ±0.70% (93 runs sampled)
scalar values (102 bytes) x 528,864 ops/sec ±0.16% (99 runs sampled)
repeated scalar fields (195 bytes) x 268,070 ops/sec ±0.16% (97 runs sampled)
map with scalar keys and values (186 bytes) x 173,328 ops/sec ±0.08% (100 runs sampled)
repeated field with 1000 messages (2000 bytes) x 2,798 ops/sec ±0.12% (101 runs sampled)
map field with 1000 messages (8890 bytes) x 1,746 ops/sec ±0.13% (99 runs sampled)
```